небольшие изменения системы очереди, фиксящие проблему, из за которой она не двигалась

+++ main.py
@@ -72,6 +72,7 @@ async def command_start_handler(message: Message) -> None:
 @dp.message()
 async def lame_message_handler(message: types.Message) -> None:
     global PRIVET_VOVAN_FLAG
+    global NUM, QUEUE, LAST_USE_TIME
 
     # id пользователя и текст его сообщения
     text = message.text
@@ -83,13 +84,15 @@ async def lame_message_handler(message: types.Message) -> None:
             if user_id == vovan and PRIVET_VOVAN_FLAG:
                 await message.reply("привет, вован")
                 PRIVET_VOVAN_FLAG = False
+
             if bot_name in text:
-                global NUM, QUEUE, LAST_USE_TIME
+                # встали в очередь
+                pos, NUM = NUM, NUM + 1
+                QUEUE += (pos,)
+                queue = None
+                generating = None
+
                 try:
-                    # встали в очередь
-                    pos, NUM = NUM, NUM + 1
-                    QUEUE += (pos,)
-                    queue = None
                     if (QUEUE.index(pos) != 0 or LAST_USE_TIME + timedelta(seconds=open_ai_kd) > datetime.utcnow()):
                         queue = await message.reply("Твой запрос в очереди на обработку!")
                     # ожидание очереди
@@ -98,13 +101,13 @@ async def lame_message_handler(message: types.Message) -> None:
                         # получаем ответ от модели
                     generating = await message.reply("генерирую ответ . . .")
                     bot_answer = await gpt_response_creation(text.replace(bot_name, 'bot'))
+                    await message.reply(bot_answer)
+                except Exception as e:
                     if queue is not None:
                         await queue.delete()
-                    await generating.delete()
-                    await message.reply(bot_answer)
-                finally:
-                    # двигаем очередь
-                    QUEUE, LAST_USE_TIME = QUEUE[1:], datetime.utcnow()
+                    if generating is not None:
+                        await generating.delete()
+                    await message.reply("че т не идет((((")
 
         elif chat_id == test:
             await message.answer("подтвердите пересылку в другой чат в консоли (надо энтер нажать)")
@@ -121,6 +124,10 @@ async def lame_message_handler(message: types.Message) -> None:
 
     except Exception as e:
         print(f"тварь на {message.from_user.id} / {message.from_user.full_name} че то сломала:\n{e}")
+    finally:
+        if bot_name in text:
+            # двигаем очередь
+            QUEUE, LAST_USE_TIME = QUEUE[1:], datetime.utcnow()
 
 
 async def main() -> None:
